Move EdgeLUT.
